fix(pricing): render faq answers as components
